+++ src/test/java/com/ejoongseok/wmslive/location/fature/Location.java
@@ -1,0 +1,33 @@
+package com.ejoongseok.wmslive.location.fature;
+
+import org.springframework.util.Assert;
+
+class Location {
+    private final String locationBarcode;
+    private final StorageType storageType;
+    private final UsagePurpose usagePurpose;
+    @lombok.Getter
+    private Long locationNo;
+
+    public Location(
+            final String locationBarcode,
+            final StorageType storageType,
+            final UsagePurpose usagePurpose) {
+        validateConstructor(locationBarcode, storageType, usagePurpose);
+        this.locationBarcode = locationBarcode;
+        this.storageType = storageType;
+        this.usagePurpose = usagePurpose;
+    }
+
+    private void validateConstructor(final String locationBarcode, final StorageType storageType, final UsagePurpose usagePurpose) {
+        Assert.hasText(locationBarcode, "로케이션 바코드는 필수입니다.");
+        Assert.notNull(storageType, "보관 타입은 필수입니다.");
+        Assert.notNull(usagePurpose, "보관 목적은 필수입니다.");
+    }
+
+    public void assignNo(final Long locationNo) {
+        this.locationNo = locationNo;
+    }
+
+
+}
