+++ src/test/java/com/ejoongseok/wmslive/location/fature/StorageType.java
@@ -1,0 +1,10 @@
+package com.ejoongseok.wmslive.location.fature;
+
+enum StorageType {
+    TOTE("토트 바구니");
+    private final String description;
+
+    StorageType(final String description) {
+        this.description = description;
+    }
+}
