+++ src/test/java/com/ejoongseok/wmslive/location/fature/UsagePurpose.java
@@ -1,0 +1,10 @@
+package com.ejoongseok.wmslive.location.fature;
+
+enum UsagePurpose {
+    MOVE("이동 목적");
+    private final String description;
+
+    UsagePurpose(final String description) {
+        this.description = description;
+    }
+}
